\name{PANGEA.BetweenHostEvolutionaryRateModifier.create.sampler.v1}
\alias{PANGEA.BetweenHostEvolutionaryRateModifier.create.sampler.v1}
\title{Create sampler of the Between Host Evolutionary Rate Multipliers}
\usage{
  PANGEA.BetweenHostEvolutionaryRateModifier.create.sampler.v1()
}
\value{
  R function
}
\description{
  Returns a function to draw the between host evolutionary
  rate multipliers.
}

