\name{PANGEA.WithinHostEvolutionaryRate.create.sampler.v1}
\alias{PANGEA.WithinHostEvolutionaryRate.create.sampler.v1}
\title{Create sampler of Within Host Evolutionary Rates}
\usage{
  PANGEA.WithinHostEvolutionaryRate.create.sampler.v1()
}
\value{
  R function
}
\description{
  Returns a function to draw within host evolutionary
  rates.
}

